requirements.txt
streamlit==1.31.0
google-cloud-geminidataanalytics==0.1.0
google-auth==2.27.0
